Speed up inference on Mac by ~3x
Changes:
+ In inference path, invoke neural net once on a batch of input
  + Also avoid expensive copies in featurizeND4j() by switching to
    featurize() then slicing out the last label in bulk
+ Change neural network input to all floats (standard in community)

This allowed my laptop (Mac pro with AVX-2 only) to match performance of a high-end server CPU
(3.0 GHz Intel Xeon Platinum 8000-series with AVX-512).

Laptop, before this change, 20/20:
+ Planning latency: {learning=14.39366491150442}
+ Planning latency: {learning=14.921567079646016}

Laptop, this change but invoke the old TDMerge(), 20/20:
+ Planning latency: {learning=45.79047076991148}
+ Planning latency: {learning=46.144710274336276}

--- To show quality is the same:

(EC2) Quality for 20/20; After (TDMV & all opts):
Improvement: {learning=15.4748217778106, left-deep=15.4748217778106, quickpick-1000=17.35624017722146}
Planning latency: {learning=11.464194350000001, left-deep=5.482218899999999, quickpick-1000=18.7576865}

(EC2) Quality for 20/20; Before (git stash):
Improvement: {learning=15.4748217778106, left-deep=15.4748217778106, nopt=-Infinity, quickpick-1000=17.35624017722146}
Planning latency: {learning=13.776011500000001, left-deep=7.5223848, nopt=3.712500000000001E-4, quickpick-1000=18.0923508}

+++ src/main/java/edu/berkeley/riselab/rlqopt/cost/CostModel.java
@@ -35,7 +35,8 @@ public interface CostModel {
     System.out.println(treeMap);
   }
 
-  default double cardinality(Attribute a) {
-    return 1.0;
+  default float cardinality(Attribute a) {
+    return 1.0f;
   }
+
 }
