+++ src/main/java/edu/berkeley/riselab/rlqopt/opt/learning/TrainingDataGenerator.java
@@ -88,18 +88,18 @@ public class TrainingDataGenerator {
     int p = 0;
 
     for (TrainingDataPoint tr : planner.getTrainingData()) {
-      Double[] vector = tr.featurize(db, c);
+      float[] vector = tr.featurize(db, c);
       p = vector.length;
 
       float[] xBuffer = new float[p - 1];
 
-      for (int ind = 0; ind < vector.length - 1; ind++) xBuffer[ind] = vector[ind].floatValue();
+      for (int ind = 0; ind < vector.length - 1; ind++) xBuffer[ind] = vector[ind];
 
       float[] yBuffer = new float[1];
 
-      if (Double.isInfinite(vector[vector.length - 1].floatValue())) continue;
+      if (Double.isInfinite(vector[vector.length - 1])) continue;
 
-      yBuffer[0] = vector[vector.length - 1].floatValue();
+      yBuffer[0] = vector[vector.length - 1];
 
       trainingExamples.add(Nd4j.create(xBuffer, new int[] {1, p - 1}));
       reward.add(Nd4j.create(yBuffer, new int[] {1, 1}));
